ai-service/requirements.txt
grpcio
grpcio-tools
# vllm
# faster-whisper
nats-py
pydantic
